Resolving lintis erorrs when runing npm run build
